Change CurrentOrderSummary.selection_id type to int
Previously CurrentOrderSummary.selection_id was declared to be a float, however the
API documentation says that it is a long:

https://api.developer.betfair.com/services/webapps/docs/x/SIA6#BettingTypeDefinitions-CurrentOrderSummary

This changes the declared data type in the model to int. int is used
over long as this maintains Python 3 compatability (in Python 2
int('<some number>') will return a long if the number is big enough
to require it).

+++ betfair/models.py
@@ -287,7 +287,7 @@ class UpdateInstruction(BetfairModel):
 class CurrentOrderSummary(BetfairModel):
     bet_id = Field(DataType(six.text_type), required=True)
     market_id = Field(DataType(six.text_type), required=True)
-    selection_id = Field(DataType(float), required=True)
+    selection_id = Field(DataType(int), required=True)
     handicap = Field(DataType(float), required=True)
     price_size = Field(DataType(PriceSize), required=True)
     bsp_liability = Field(DataType(float), required=True)
